Final refactoring for compliance. Now BHoM compliant. Needs testing.
