+++ TDRepo_Adapter/TDRepo_Adapter.csproj
@@ -125,8 +125,12 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AdapterActions\Push.cs" />
+    <Compile Include="Convert\AdapterId.cs" />
+    <Compile Include="Convert\ToTDRepo\Mesh.cs" />
     <Compile Include="Convert\ToTDRepo\MeshToGeom.cs" />
+    <Compile Include="Extra\Commit.cs" />
     <Compile Include="Extra\UploadBIM.cs" />
+    <Compile Include="Extra\WriteOBJFile.cs" />
     <Compile Include="Extra\UploadOBJ.cs" />
     <Compile Include="AdapterActions\Execute.cs" />
     <Compile Include="Extra\WriteBIMFile.cs" />
