+++ TDRepo_Engine/Compute/MeshRepresentation/3DElements/Sphere.cs
@@ -36,7 +36,6 @@ using BH.oM.Structure.Elements;
 using BH.Engine.Structure;
 using BH.Engine.Rhinoceros;
 using BH.oM.Structure.Constraints;
-using BH.oM.External.TDRepo;
 using BH.oM.Analytical.Elements;
 
 namespace BH.Engine.External.TDRepo
@@ -44,14 +43,14 @@ namespace BH.Engine.External.TDRepo
     public static partial class Compute
     {
         [Description("Returns a BHoM mesh representation for the BHoM Bar.")]
-        public static BH.oM.Geometry.Mesh MeshRepresentation(this Sphere sphere, DisplayOptions displayOptions = null)
+        public static BH.oM.Geometry.Mesh MeshRepresentation(this Sphere sphere, BH.oM.External.TDRepo.DisplayOptions displayOptions = null)
         {
-            displayOptions = displayOptions ?? new DisplayOptions();
+            displayOptions = displayOptions ?? new BH.oM.External.TDRepo.DisplayOptions();
 
             return sphere.RhinoMeshRepresentation(displayOptions).FromRhino();
         }
 
-        public static Rhino.Geometry.Mesh RhinoMeshRepresentation(this BH.oM.Geometry.Sphere sphere, DisplayOptions displayOptions)
+        public static Rhino.Geometry.Mesh RhinoMeshRepresentation(this BH.oM.Geometry.Sphere sphere, BH.oM.External.TDRepo.DisplayOptions displayOptions)
         {
             return Rhino.Geometry.Mesh.CreateFromSphere(sphere.ToRhino(), 8, 4);
         }
