+++ TDRepo_oM/Mesh.cs
@@ -26,7 +26,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.oM.TDRepo
+namespace BH.oM.External.TDRepo
 {
     public class Mesh
     {
